+++ .env.example
@@ -1,5 +1,3 @@
-PINATA_KEY=1234
-PINATA_SECRET=1234
 PROD=false
 NEXT_PUBLIC_INFURA_ID=1234
 NEXT_PUBLIC_WIKI_CONTRACT_ADDRESS=0x94bb4c72252d0ae7a98b2b0483Dc4145C0C79059
