+++ package.json
@@ -30,8 +30,8 @@
     "@emotion/react": "^11.8.2",
     "@emotion/styled": "^11.8.1",
     "@ensdomains/ens-avatar": "^0.1.2",
-    "@ethersproject/bignumber": "^5.5.0",
-    "@ethersproject/providers": "^5.5.2",
+    "@ethersproject/bignumber": "^5.6.2",
+    "@ethersproject/providers": "^5.6.8",
     "@everipedia/wagmi-magic-connector": "0.3.6",
     "@magic-ext/oauth": "^2.1.1",
     "@magic-sdk/provider": "^8.1.1",
@@ -51,7 +51,7 @@
     "chakra-ui-steps": "^1.7.2",
     "date-fns": "^2.28.0",
     "debounce": "^1.2.1",
-    "ethers": "5.6.2",
+    "ethers": "^5.6.8",
     "fast-diff": "^1.2.0",
     "framer-motion": "^5.6.0",
     "fuse.js": "^6.6.2",
@@ -80,7 +80,7 @@
     "remark-gfm": "^3.0.1",
     "slick-carousel": "^1.8.1",
     "slugify": "^1.6.5",
-    "wagmi": "^0.3.5",
+    "wagmi": "^0.4.3",
     "walletlink": "^2.5.0",
     "yarn": "^1.22.18"
   },
